:recycle: Refactor Code - Exercício 02
