+++ nt-exercises/src/week02/exercise02/Main.java
@@ -1,8 +1,6 @@
-package week02.exercise02.test;
+package week02.exercise02;
 
-import week02.exercise02.classes.CaesarCipher;
-
-public class CaesarCipherTest {
+public class Main {
     public static void main(String[] args) {
         CaesarCipher caesarCipher = new CaesarCipher();
         // Exemplo: Java é uma linguagem incrível
