+++ nt-exercises/src/week02/exercise02/utils/EncoderDecoder.java
@@ -1,6 +1,6 @@
 package week02.exercise02.utils;
 
-import week02.exercise02.classes.CaesarCipher;
+import week02.exercise02.CaesarCipher;
 
 import java.text.Normalizer;
 
@@ -10,9 +10,9 @@ public class EncoderDecoder extends CaesarCipher {
         String messageOrWordNormalized = Normalizer.normalize(messageOrWord, Normalizer.Form.NFD).replaceAll("[^\\p{ASCII}]", "");
 
         for (char c : messageOrWordNormalized.toCharArray()) {
+            char baseMessageOrWord = Character.isLowerCase(c) ? 'a' : 'A';
+            char encryptedChar = (char) (baseMessageOrWord + (c - baseMessageOrWord + change) % 26);
             if (Character.isLetter(c)) {
-                char baseMessageOrWord = Character.isLowerCase(c) ? 'a' : 'A';
-                char encryptedChar = (char) (baseMessageOrWord + (c - baseMessageOrWord + change) % 26);
                 encryptedMessageOrWord.append(encryptedChar);
             } else {
                 encryptedMessageOrWord.append(c);
